quic: move the RXE definition to a local header
The RXE structure carries the reference count of a received packet and
embeds the OSSL_QRX_PKT handed out to the users of the QRX as its first
member. It is defined in quic_record_rx.c, so a test that wants to build
a packet without a QRX behind it cannot allocate one or look at its
reference count, and would have to guess the size of the structure and
the offset of the fields it needs.

Move the definition to a local header, following the other local headers
in ssl/quic, so that a test can include it and construct a packet of its
own rather than the library having to provide a constructor for it.


MergeDate: Fri Aug 28 14:03:45 2026
(Merged from https://github.com/openssl/openssl/pull/32515)
